Update dataplane.yml - protocol values
The default cifv5 sqlite schema creates protocol as type INTEGER in the indicators table. There is no code that normalizes these text values for protocol to integers, and this appears to be the only default feed config that uses the protocol key, so the easy fix to avoid type errors (and rollbacks) from sqlite3 is to convert the protocols to integers here.